Reorganise address controller
