--- app/policies/team_members/address_policy.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-class TeamMembers::AddressPolicy < ApplicationPolicy
-  def show?
-    user.has_any_role?(
-      { name: :admin, resource: record.company },
-      { name: :owner, resource: record.company }) || user == record.user
-  end
-
-  def update?
-    user.has_any_role?(
-      { name: :admin, resource: record.company },
-      { name: :owner, resource: record.company }) || user == record.user
-  end
-end
