+++ app/views/internal_api/v1/addresses/_address.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.extract! address, :address_type, :address_line_1, :address_line_2, :city, :state, :country, :pin
