+++ app/views/internal_api/v1/addresses/update.json.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.partial! "address", locals: { address: }
