--- app/views/internal_api/v1/team_members/addresses/_address.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! user.addresses, :address_type, :address_line_1, :address_line_2, :city, :state, :country, :pin
