--- app/views/internal_api/v1/team_members/addresses/update.json.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.partial! "address", locals: { user: }
