+++ config/routes/internal_api.rb
@@ -72,8 +72,8 @@ namespace :internal_api, defaults: { format: "json" } do
 
     resources :team, only: [:index, :destroy] do
       resource :details, only: [:show, :update], controller: "team_members/details"
-      resource :addresses, only: [:show, :update], controller: "team_members/addresses"
     end
+    resources :addresses, only: [:show, :update], controller: "addresses"
 
     resource :profile, only: [:update, :show], controller: "profile" do
       delete "/remove_avatar", to: "profile#remove_avatar"
